Update README.md
I am unable to connect github to dbt as there is error in setting up repository in dbt. I was not able to create visualizations using google-visual-studio due to this problem. As instructed by mentor, I have written the code in snowflake warehouse and pasted it here by creating a new repository